requirements/dev.txt
-r requirements.txt

pylint==1.4.1
pylint-plugin-utils==0.2.3
vcrpy==1.4.1
